+++ es5/index.html
@@ -34,6 +34,7 @@
       <span class="versions">
         <a href="../es5" class="current">5</a><!--
         --><a href="../es6">6</a><!--
+        --><a href="../es2016plus">ES2016+</a><!--
         --><a href="../esnext">next</a><!--
         --><a href="../esintl">intl</a><!--
         --><a href="../non-standard">non-standard</a>
