+++ esintl/index.html
@@ -49,6 +49,7 @@
       <span class="versions">
         <a href="../es5">5</a><!--
         --><a href="../es6">6</a><!--
+        --><a href="../es2016plus">ES2016+</a><!--
         --><a href="../esnext">next</a><!--
         --><a href="../esintl" class="current">intl</a><!--
         --><a href="../non-standard">non-standard</a>
