+++ esnext/skeleton.html
@@ -31,6 +31,7 @@
       <span class="versions">
         <a href="../es5">5</a><!--
         --><a href="../es6">6</a><!--
+        --><a href="../es2016plus">ES2016+</a><!--
         --><a href="../esnext" class="current">next</a><!--
         --><a href="../esintl">intl</a><!--
         --><a href="../non-standard">non-standard</a>
