+++ non-standard/index.html
@@ -33,6 +33,7 @@
         <span class="versions">
           <a href="../es5">5</a><!--
           --><a href="../es6">6</a><!--
+          --><a href="../es2016plus">ES2016+</a><!--
           --><a href="../esnext">next</a><!--
           --><a href="../esintl">intl</a><!--
           --><a href="../non-standard" class="current">non-standard</a>
